Rename complete to suggestions for consistency
