--- frontend/src/metabase/querying/expressions/complete/index.ts
@@ -1,3 +1,0 @@
-export * from "./complete";
-export { expressionClauseSnippet } from "./util";
-export { type Completion } from "./types";
