+++ frontend/src/metabase/querying/expressions/index.ts
@@ -1,9 +1,9 @@
 export * from "./clause";
-export * from "./complete";
 export * from "./diagnostics";
 export * from "./errors";
 export * from "./formatter";
 export * from "./help-text";
 export * from "./mode";
 export * from "./position";
+export * from "./suggestions";
 export * from "./types";
