+++ frontend/src/metabase/querying/expressions/suggestions/functions.ts
