+++ frontend/src/metabase/querying/expressions/suggestions/index.ts
@@ -1,0 +1,2 @@
+export * from "./suggest";
+export { type ExpressionSuggestion } from "./types";
