+++ frontend/src/metabase/querying/expressions/suggestions/literals.ts
